End point faltantes y primeras validaciones

+++ src/main/java/com/example/seguimientocuerpos/controller/MinutaController.java
@@ -2,7 +2,10 @@ package com.example.seguimientocuerpos.controller;
 
 import com.example.seguimientocuerpos.model.Minuta;
 import com.example.seguimientocuerpos.service.MinutaService;
+import com.example.seguimientocuerpos.service.UsuarioService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,6 +16,7 @@ import java.util.List;
 public class MinutaController {
 
     private final MinutaService minutaService;
+    private final UsuarioService usuarioService;
 
     @PostMapping("/minuta")
     public void save(@RequestBody Minuta minuta){
@@ -29,8 +33,12 @@ public class MinutaController {
         return minutaService.findById(id).get();
     }
 
-    @DeleteMapping("/minuta/{id}")
-    public void deleteById (@PathVariable String id){
-        minutaService.deleteById(id);
+    @DeleteMapping("/minuta/{id}/{id_usuario}")
+    public ResponseEntity<Object> deleteById (@PathVariable String id, @PathVariable String id_usuario){
+        if(usuarioService.findById(id_usuario).get().getRol().equals("LIDER")) {
+            minutaService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 }
