+++ src/main/java/com/example/seguimientocuerpos/controller/ProduccionAcademicaController.java
@@ -3,6 +3,8 @@ package com.example.seguimientocuerpos.controller;
 import com.example.seguimientocuerpos.model.ProduccionAcademica;
 import com.example.seguimientocuerpos.service.ProduccionAcademicaService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -29,13 +31,21 @@ public class ProduccionAcademicaController {
         return produccionAcademicaService.findById(id).get();
     }
 
-    @DeleteMapping("/produccionAcademica/{id}")
-    public void deleteById(@PathVariable String id){
-        produccionAcademicaService.deleteById(id);
+    @DeleteMapping("/produccionAcademica/{id}/{id_usuario}")
+    public ResponseEntity<Object> deleteById(@PathVariable String id, @PathVariable String id_usuario){
+        if(produccionAcademicaService.findById(id).get().getId_usuario().equals(id_usuario)) {
+            produccionAcademicaService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PutMapping("/produccionAcademica")
-    public void update(@RequestBody ProduccionAcademica produccionAcademica){
-        produccionAcademicaService.save(produccionAcademica);
+    @PutMapping("/produccionAcademica/{id_usuario}")
+    public ResponseEntity<Object> update(@RequestBody ProduccionAcademica produccionAcademica, @PathVariable String id_usuario){
+        if(produccionAcademicaService.findById(produccionAcademica.getId()).get().getId_usuario().equals(id_usuario)) {
+            produccionAcademicaService.save(produccionAcademica);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 }
